refactor(channels): split WhatsApp messaging out of meta.py
meta.py had grown to 774 lines across four responsibilities. Onboarding an
account and sending on an already-connected one are different jobs done by
different roles — org admins vs inbox agents — and only the latter cares about
WhatsApp Business Accounts and templates.

- New whatsapp_messaging.py holds the template and outbound-conversation
  routes with the helpers only they use (_whatsapp_account_or_404,
  _fetch_all_templates, _waba_credentials, require_inbox_agent).
- graph_detail was the one helper both halves need, so it moves to meta_base
  beside the Graph transport whose error bodies it reads, rather than being
  imported privately across API modules.
- Both routers mount under /meta, so every URL is unchanged.

meta.py 774 -> 599 lines. No behaviour change: the same 171 tests pass, with
only patch targets repointed at the moved code.
